+++ clients/algoliasearch-client-csharp/algoliasearch/Http/Request.cs
@@ -22,7 +22,7 @@ public class Request
   public Uri Uri { get; set; }
 
   /// <summary>
-  /// Headers a dictionary
+  /// Headers 
   /// </summary>
   public IDictionary<string, string> Headers { get; set; }
