+++ clients/algoliasearch-client-csharp/algoliasearch/Models/Common/AbstractSchema.cs
@@ -11,16 +11,6 @@ public abstract class AbstractSchema
   public abstract object ActualInstance { get; set; }
 
   /// <summary>
-  /// Gets or Sets IsNullable to indicate whether the instance is nullable
-  /// </summary>
-  public bool IsNullable { get; protected set; }
-
-  /// <summary>
-  /// Gets or Sets the schema type, which can be either `oneOf` or `anyOf`
-  /// </summary>
-  public string SchemaType { get; protected set; }
-
-  /// <summary>
   /// Converts the instance into JSON string.
   /// </summary>
   public abstract string ToJson();
